pyproject.toml
[project]
name = "chunking4rag"
authors = [{ name = "Harpreet Sethi", email = "harpreetset@gmail.com" }]
version = "0.0.4"
description = "A small library to chunk large files into smaller arrays that can be used for generating RAG embeddings"
readme = "README.md"
classifiers = [
    "Development Status :: 3 - Alpha",
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
]
requires-python = ">=3.12"
dependencies = [
    "fastexcel>=0.13.0",
    "gensim>=4.3.3",
    "html2text>=2024.2.26",
    "nltk>=3.9.1",
    "pillow>=11.1.0",
    "polars>=1.23.0",
    "pydantic>=2.10.6",
    "pypdf>=5.3.0",
    "pytesseract>=0.3.13",
]
[project.optional-dependencies]
test = [
    "pytest",
    "pytest-cov",
]
formatting = [
    "black",
    "flake8",
    "isort",
]
type-checking = [
    "mypy",
]
docs = [
    "pydocstyle",
]
publishing = [
    "twine",
    "wheel"
]
[build-system]
requires = ["setuptools", "wheel"]
build-backend = "setuptools.build_meta"

[dependency-groups]
dev = [
    "chunking4rag",
]

[tool.uv.sources]
chunking4rag = { workspace = true }
